selftests/livepatch: Create a simple selftest for state callbacks
The per-object callbacks have been deprecated in favor of per-state
callbacks and will be removed soon. The behavior of these two types
of callbacks is completely different.

Per-object callbacks are triggered when the livepatching of
the related object is enabled or disabled. When using the atomic
replace feature, only the callbacks from the new livepatch are called.

Per-state callbacks, on the other hand, are triggered when the state
is being added or removed. When using the atomic replace feature,
the callbacks that remove the state are called from the old livepatch
that is being replaced.

For testing purposes, the state represents a part of an audience in
the livepatch for the speaker module. The audience applauds when
the speaker delivers the welcome message.

All four callbacks are used as follows:

  + pre_patch() allocates a shadow variable with a string and fills
		it with "[]".
  + post_patch() fills the string with "[APPLAUSE]".
  + pre_unpatch() reverts the string back to "[]".
  + post_unpatch() releases the shadow variable.

The welcome message printed by the livepatched function allows us to
distinguish between the transition and the completed transition.
Specifically, the speaker's welcome message appears as:

  + Not patched system:		 "Hello, World!"
  + Transition (unpatched task): "[] Hello, World!"
  + Transition (patched task):	 "[] Ladies and gentlemen, ..."
  + Patched system:		 "[APPLAUSE] Ladies and gentlemen, ..."

Create an initial self-test that uses a single livepatch and
a single state. This serves as a base for more complex self-tests.

Signed-off-by: Petr Mladek <pmladek@suse.com>