samples/livepatch: Replace sample module with obsolete per-object callbacks
The per-object callbacks have been deprecated in favor of per-state
callbacks and are scheduled for removal.

Remove the sample modules with the obsolete per-object callbacks.
Instead add new sample modules based on the new selftests for
per-state callbacks.

The test modules are just renamed to follow the naming scheme of
the existing sample modules. Also only one version of the test
module and one version of the livepatch module is built.

Signed-off-by: Petr Mladek <pmladek@suse.com>