selftests/livepatch: Move basic tests for livepatching modules
The per-object callbacks have been replaced by per-state callbacks and
will be removed soon.

There are self-tests that attempt to load a live patch with per-object
callbacks and the live-patched module (object) in all possible order
variations.

These order variations are less relevant with per-state callbacks because
they are only invoked when the live patch is enabled or disabled.

However, it is still important to check all possible order variations
to ensure the module is live-patched correctly. The expected state is
validated using two approaches.

First, the "patched" sysfs attribute is checked the same way is in
the sysfs tests.

Second, a new "welcome" parameter has been added to the speaker test
module. By reading this parameter's state, the live-patched function can
be invoked.

Instead of displaying the parameter value via the sysfs interface, the
.read() callback writes a welcome message to the kernel log. This message
is compared with the expected value at the end of the test. This approach
ensures that the self-test attempts to unload the test modules even if
something goes wrong.

Signed-off-by: Petr Mladek <pmladek@suse.com>