selftests/livepatch: Convert testing of multiple target modules
The per-object callbacks have been replaced by per-state callbacks and
will be removed soon.

A selftest previously loaded two livepatched modules to ensure that
the per-object callbacks were called for both.

Although per-state callbacks are only invoked when the live patch is
enabled or disabled, it is still important to verify that multiple objects
can be live-patched.

Convert the test into a generic one by reusing the speaker test module
to create the second target module.

In the second variant, speaker_welcome() will only print "(2)" after
the function name to distinguish it, while keeping the real function name
the same to maintain simplicity.

Signed-off-by: Petr Mladek <pmladek@suse.com>